MakeCoMaster:
We allow breaking of an existing co-master replication. Here's the
breakdown:
Ideally, this would not eb allowed, and we would first require the user
to RESET SLAVE on 'master'
prior to making it participate as co-master with our 'instance'.
However there's the problem that upon RESET SLAVE we lose the
replication's user/password info.
Thus, we come up with the following rule:
If S replicates from M1, and M1<->M2 are co masters, we allow S to
become co-master of M1 (S<->M1) if:

- M1 is writeable
- M2 is read-only or is unreachable/invalid
- S  is read-only

And so we will be replacing one read-only co-master with another.
(instance_topology.go)

supported by Web interface (cluster.js)