Allow web app to work through URL-rewriting proxies.
Specifically, this lets it work when accessed through the Kubernetes
apiserver proxy, which looks like this to your browser:

http://localhost/api/v1/proxy/.../orchestrator/...

The apiserver proxy rewrites absolute URLs where it can, in HTML and in
HTTP headers, but URLs generated from client-side code (JS) need to be
considerate of the fact that the base URL of the app might not be at the
root of the host.

One solution would be to make all URLs relative, but that gets messy for
apps like Orchestrator that use various levels of path nesting:
  /web/clusters
  /web/cluster/alias/:clusterAlias

Instead, we generate absolute URLs after looking (on the client side) to
see where the user's browser thinks the app is based. This is based on
the assumption that the right-most occurrence of "/web/" in the browser
path indicates the base of the Orchestrator web app.