calling on a nil value is an error
We could return err when called on a uninitialized value
but silently skipping execution does not seem right to me.

I think we should not (yet) guard against such programming
errors (and calling HandleRequests on nil is an error)
because 1) the module is not meant to be used frequently
and 2) there're not too many contributors yet.

So let's save a few lines.