splitting discovery queue and worker pool in two separate parts
IMO it's easier to understand and modify this way.
Discovery queue becomes a single channel with a deduplication
logic. A worker pool is a bunch of simple goroutines,
consuming this queue.

For simplicity, we create all workers on start. By default,
it's 300 goroutines. The footprint used to be 4K per
gouroutine, thus should not be a concern. An CPU cost
of idle gouroutines is pretty much zero.

I'm also taking many debug statements out to prepare
for upstream.