use readTimeout in mysql driver and relevant config option
We had a couple of misbehaving DB machines in our environment.
The kernel was able to accept and serve tcp connections
but mysqld was unresponsive. As a result, plenty of discovery
goroutines in Orchestrator would hang forever on sql.Query
calls.

We've chosen 10s as a default read timeout for both
topology and backend reads. We think 10 seconds is a
safe bet: If db hosts or backend is not responding
that long then there must be a problem that should
be handled specially.