optimize how last_attempted_check is updated
On success last_attempted_check is updated by writeInstance anyway,
so there's no need to update it twice. Most of the time
we get success. Thus this update essentially halfs
the number of write queries on backend db.

We make sure that UpdateInstanceLastAttemptedCheck() is called
on every err return from the ReadTopologyInstance function.