do more logging when running shell commands
Shell hooks are typically executed on failures/recoveries.  A
failure to execute a hook is critical and it's good to log it
verbosely. Such events does not happen often, thus this does not
affect log size much.

Also added a new parameter: ProcessesShellCommand to allow
configuration of which shell to use. Configured to "bash"
by default for backward compatibility. This allows using
absolute paths to shells when people find it convenient.