Update Dockerfile
Just reduced the size of the docker image from 3xxMB to 40MB.
Started from importing from a smaller image, plain alpine, then installed go and build-base to compile the project.
All the rest is almost the same as the original, just opted to store GOPATH under /tmp and also delete it at the end of the build.

I hope it helps. Keep up the great work!
Orchestrator is a hell of a tool!