NodeHealthExpiry: if false ensure the expiry happens on the active node
All nodes, "cli clients" or active "http" nodes, insert into the
node_health table, and keep it up to date while they are running.
By default they also try to clean up old entries in this table. If
a large number of http or cli nodes are running then this clean up
may trigger unwanted contention on the table, so it has been possible
for some time to disable the expiry. However, this requires a clean
up, and the best node to do this is the active orchestrator node.
This patch thus checks if NodeHealthExpiry is false and if so the
active node will carry out this task.