make the configuration race detector safe.
The go race detector notices that we may overwrite the configuration
while reading it as there are no locks to prevent this.  This is
most likely to happen when you send SIGHUP to the running process
as it is quite busy.

So replace config.Config with config.Config() and return a copy of
the pointer to the configuration.

Load() Reload() and ForceReload() have been adjusted to use locking
and to also create a new copy of the configuration rather than
modify the existing one. The new copy then replaces the original
configuration and this process is surrounded by locking.  Configuration
modifications by the application are now done via calls which also
lock.