Add govendor support for github.com/martini-contrib/render
While debugging something I wanted to check if this was the most
up to date code and currently that info was not available. Hence
added the info.  I ran:

$ govendor fetch "github.com/martini-contrib/render"

which pulled in the latest changes (minimal in this case so no
change) but at least now vendor.json includes the commit that was
used from this vendor software.