Adjust db.SetMaxIdleConns for backend orchestrator to a more dynamic value
Recent issues seen talking to the backend can give a: cannot assign
requested address error. The number of open tcp sockets to the backend
went over the limit.  This was triggered by the hard-coded value of
db.SetMaxIdleConns(10) intended to prevent too many backend database
connections being left open triggering the go driver to close connections
and shortly afterwards open them again triggering the issue described.

This patch now adjusts the value of SetMaxIdleConns to 25% of
MySQLOrchestratorMaxPoolConnections if this value is larger and so will
trigger less reconnections and help aleviate this point of stress.

I chose not to make the setting configurable (yet) as this may confuse
people. If you make MySQLOrchestratorMaxPoolConnections higher then
the number of idle connections is allowed to grow.