Add DiscoveryIgnoreHostnameFilters and deprecate DiscoveryIgnoreReplicaHostnameFilters
It may be necessary to filter out accessing the master as it may not be
reachable (e.g. if in an external network) so this change changes the
use of DiscoveryIgnoreReplicaHostnameFilters to the more general name
DiscoveryIgnoreHostnameFilters.

Use of the old variable name will still work but will generate a
deprecation warning.

Additional extra checks are added to ensure that a host that we want to
ignore really is ignored (the config may change on a running orchestrator)