Fix vagrant networking configuration
With a fresh install of Vagrant 2.0.1, and
VirtualBox 5.2.2r119230 on OSX the vagrant config
failed.  In particular the host machine was unable
to connect via ssh because the network to the VM
was unreachable.

I changed some of the extra config options passed
to VirtualBox for network config and it solved the issue.

In addition there were a some path issues -- it looks like
the build.sh script has changed and needs to be able to
run a `git revlist` to build the package.