Remove local ssh key file setup
This part of the vagrant build referenced
key files in the project tree which didn't exist.
In order to keep setup as simple as possible it was
easier to remove this part instead of either adding
dummy keys to the repo or changing the instructions
to ask the user to generate them.