Fix MySQL backend TLS/SSL configuration.
Setting `MySQLOrchestratorUseMutualTLS=false` prevented _any_ SSL
connection to the backend server, as it guarded the call to
`SetupMySQLOrchestratorTLS`. This behaviour is inconsistent with the
documentation. To keep full backwards compatibility but also correct
this bug, a value set in `MySQLOrchestratorSSLCAFile` will now result in
connections that use SSL.

Further, unlike the topology code path, `MySQLOrchestratorUseMutualTLS`
was not referenced within `SetupMySQLOrchestratorTLS`. As such, all
connections required a (valid) client key pair to be provided,
preventing the application from starting. This change unifies the code
paths.