Display region, data center, and environment in UI
The region, data center, and physical environment bits of metadata are not currently displayed in the UI, though instances have a colored outline/border to distinguish DCs. This adds all 3 pieces of information to the instance modal window and also the title/tooltip for each instance in the cluster screen. (The latter has been tested on FF, Chrome, and Safari on OS X, but the embedded new lines may not work in all browsers.) This also adds a legend to the bottom of the cluster screen that shows the mapping of outline color to DC.

Fixes #1094.